Only feed in list of unique texts to model
If we have 10 frames in a video, we do not feed in the same text description 10 times during testing. This makes testing more efficient and makes the accuracy metric more sensible. The only downside is that it requires us to have two separate lists in the Dataset list: one keeping track of the text associated with every frame, and another only keeping track of the unique texts.